+++ projects/angular2-smart-table/src/lib/components/filter/filter-types/select-filter.component.ts
@@ -3,6 +3,7 @@ import {NgControl} from '@angular/forms';
 import {debounceTime, distinctUntilChanged, skip} from 'rxjs/operators';
 
 import {DefaultFilter} from './default-filter';
+import {defaultStringEqualsFilter, defaultStringInclusionFilter} from "../../../lib/data-source/local/local.filter";
 
 @Component({
   selector: 'select-filter',
@@ -23,11 +24,9 @@ export class SelectFilterComponent extends DefaultFilter implements OnInit {
 
   @ViewChild('inputControl', { read: NgControl, static: true }) inputControl!: NgControl;
 
-  constructor() {
-    super();
-  }
-
   ngOnInit() {
+    this.column.filterFunction = this.onFilterValues.bind(this);
+
     const exist = this.inputControl.valueChanges;
     if (!exist) {
       return;
@@ -39,5 +38,14 @@ export class SelectFilterComponent extends DefaultFilter implements OnInit {
         debounceTime(this.delay)
       )
       .subscribe((value: string) => this.setFilter());
+  }
+
+  onFilterValues(cellValue: string, search: string, data: any, cellName: string) {
+    const strictFilter = this.column.getFilterConfig()?.strict ?? false;
+    if (strictFilter) {
+      return defaultStringEqualsFilter(cellValue, search, data, cellName);
+    } else {
+      return defaultStringInclusionFilter(cellValue, search, data, cellName);
+    }
   }
 }